Extract distribution IDE resolver to a separate class
